CMake: guard unit-tests
Adds an option to control whether unittests should be registered with
cmakes ctest runner. It is initialized based on whether pffft is the
current top level cmake build directory.

This helps with including pffft via FetchContent with EXCLUDE_FROM_ALL.
Because cmake will only build the targets that are actually required by
the consuming project, if all tests are unconditionally registered with
ctest, the testsuite of the consuming project will fail. This assumes
that most library consumers are not interested in running the testsuites
of all their dependencies in their own build.

Also move the definition of the compile definition into the PFFFT
target, so it only has to be defined once and downstream consumers can
just include the target.